feat: add logic result

+++ app/src/main/java/com/example/bmicalculator/data/WeightProperties.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class WeightProperties(val type: String, val weight: Int, val height: Int) : Parcelable
+data class WeightProperties(val type: String, val weight: Double, val height: Double) : Parcelable
